+++ include/vb/Coloring.h
@@ -41,7 +41,7 @@ namespace vb {
     TEST_CASE("vb::Coloring") {
         char *   argv[] = {(char *)"test_coloring"};
         Hub      H("Testing Coloring", 1, argv);
-        Coloring C(H.title, cpx(0, 0), cpx(1, 1), 500, [](cpx z) { return Indexed(real(z) * imag(z) * 10); });
+        Coloring C(H.title, cpx(0, 0), cpx(1, 1), 500, [](cpx z) { return Indexed(int(real(z) * imag(z) * 10)); });
         C.scale(1.2);
         C.shift(cpx(.2, .3));
         C.show();
